Lazy create tile textures/animations
This change reduces the startup time needed by not front-loading the
creation of each tile's texture and all animation data. This benefit is
largely visible on large tilemaps/tilesets.

Another benefit of this change is that now only the textures/animations
that are needed for rendering are actually in memory. Again, for large
tilemaps this can make a pretty huge difference.